refactor(calendars): #91 move shared types to types/ and enforce code boundaries
Extracts CalendarProvider, IncomingEvent, and related types from
convex/calendars/orchestrator/types.ts into a neutral types/ directory
imported via @shared/*. Adds boundary rules to CLAUDE.md and CONTEXT.md
to prevent server/client code mixing. Fixes async adapter stubs across
all four adapters (google, ical, microsoft, native).
